+++ postcss.config.js
@@ -1,9 +1,7 @@
 module.exports = {
   plugins: {
-    "postcss-nesting": {},
-    "postcss-custom-properties": {
-      importFrom: "src/components/shared.css",
-    },
+    "postcss-import": {},
     autoprefixer: {},
+    tailwindcss: {},
   },
 };