+++ src/components/AppShell.tsx
@@ -1,19 +1,13 @@
 import React, { FunctionComponent, useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { Helmet } from 'react-helmet';
-import { Route, useHistory } from "react-router";
+import { Helmet } from "react-helmet-async";
+import { Route, Routes, useLocation } from "react-router";
 import loadable from "@loadable/component";
 import Menu from "./Menu";
 import Loading from "./Loading";
-import "./AppShell.css";
 import LanguagePicker from "./LanguagePicker";
-import Item from "./Item";
 import TestDataProvider from "../context/TestDataProvider";
 import About from "./Info";
-import us from "../img/us.svg";
-import de from "../img/de.svg";
-import es from "../img/es.svg";
-import fr from "../img/fr.svg";
 import Tracking from "./Tracking";
 import useAnalytics from "../hooks/useAnalytics";
 
@@ -31,48 +25,33 @@ const Stats = loadable(() => import("./stats/Stats"), {
 
 const AppShell: FunctionComponent = () => {
   const { t, i18n } = useTranslation();
-  const { listen } = useHistory();
+  const location = useLocation();
   const { updateConfig } = useAnalytics();
 
-  useEffect(() => listen((location) => {
+  useEffect(() => {
     updateConfig({
       anonymize_ip: true,
       page_path: location.pathname,
     });
-  }), []);
+  }, [location]);
 
   return (
-    <div id="page-wrapper">
+    <div id="page-wrapper" className="font-sans flex fixed top-0 bottom-0 left-0 right-0 bg-grey-100 flex-col-reverse md:flex-row">
       <Helmet htmlAttributes={{ lang: i18n.language }} />
       <Menu />
-      <div id="page-body">
+      <div id="page-body" className="flex flex-col flex-grow overflow-auto">
         <Tracking />
-        <header>
-          <h1>{t("app.title")}</h1>
-          <LanguagePicker>
-            <Item code="en">
-              <img src={us} alt="English" />
-              <span>English</span>
-            </Item>
-            <Item code="de">
-              <img src={de} alt="Deutsch" />
-              <span>Deutsch</span>
-            </Item>
-            <Item code="fr">
-              <img src={fr} alt="Français" />
-              <span>Français</span>
-            </Item>
-            <Item code="es">
-              <img src={es} alt="Español" />
-              <span>Español</span>
-            </Item>
-          </LanguagePicker>
+        <header className="flex bg-blue-300 shadow basis-14">
+          <h1 className="mx-4 mt-4 mb-3 text-white text-xl grow">{t("app.title")}</h1>
+          <LanguagePicker />
         </header>
         <TestDataProvider>
-          <Route path="/" component={RulesTest} exact />
-          <Route path="/rules" component={HandballRules} />
-          <Route path="/stats" component={Stats} />
-          <Route path="/about" component={About} />
+          <Routes>
+            <Route index element={<RulesTest />} />
+            <Route path="rules" element={<HandballRules />} />
+            <Route path="stats" element={<Stats />} />
+            <Route path="about" element={<About />} />
+          </Routes>
         </TestDataProvider>
       </div>
     </div>