+++ src/components/CheckBox.tsx
@@ -1,7 +1,6 @@
 import React, { FunctionComponent, KeyboardEvent } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCheck } from "@fortawesome/free-solid-svg-icons";
-import "./CheckBox.css";
 
 interface Props {
   checked: boolean;
@@ -24,7 +23,7 @@ const CheckBox: FunctionComponent<Props> = ({
 
   return (
     <div
-      className="checkbox"
+      className="checkbox relative select-none bg-white border border-blue-500 rounded w-5 h-5 box-border cursor-pointer hover:bg-blue-200 aria-checked:bg-blue-400"
       role="checkbox"
       aria-checked={checked}
       tabIndex={0}
@@ -33,12 +32,13 @@ const CheckBox: FunctionComponent<Props> = ({
     >
       <input
         type="checkbox"
+        className="absolute opacity-0 w-0 h-0 peer"
         checked={checked}
         onChange={onChange}
         readOnly={readOnly}
         aria-labelledby={labelledBy}
       />
-      <FontAwesomeIcon className="checkmark" icon={faCheck} />
+      <FontAwesomeIcon className="hidden absolute text-white pl-0.5 peer-checked:block" icon={faCheck} />
     </div>
   );
 };