+++ src/components/Info.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import "./Info.css";
 import CheckBox from "./CheckBox";
 import usePrivacySettings from "../hooks/usePrivacySettings";
 import { Decision } from "../context/PrivacyContext";
+import Box from "./Box";
 
 const Info = () => {
   const {
@@ -29,33 +29,33 @@ const Info = () => {
   };
 
   return (
-    <div id="info-wrapper" lang="en">
-      <div id="info" className="info-box">
+    <div className="overflow-auto" lang="en">
+      <Box className="text-center">
         <div>Website built by</div>
-        <div id="info-name">Christoph Kraemer</div>
-        <div id="info-title">USA Team Handball</div>
-        <div id="info-mail">Contact: hello@referee.app</div>
-      </div>
-      <div id="privacy-settings" className="info-box">
-        <h2>Privacy Settings</h2>
-        <div className="privacy-header">
+        <div className="font-bold mt-2">Christoph Kraemer</div>
+        <div className="text-xs">USA Team Handball</div>
+        <div className="mt-4">Contact: hello@referee.app</div>
+      </Box>
+      <Box>
+        <h2 className="text-xl font-bold">Privacy Settings</h2>
+        <div className="flex mt-4 mb-2">
           <CheckBox
             checked={allowTracking === Decision.ACCEPTED}
             onChange={handleTrackingChanged}
           />
-          <h3>Analytics</h3>
+          <h3 className="font-bold ml-2">Analytics</h3>
         </div>
         <div>
           We use Google Analytics to analyze traffic and optimize this website based on the data.
           We also collect data on which rules test questions were answered and if the answer was
           correct or wrong. More information can be found in the privacy policy below.
         </div>
-        <div className="privacy-header">
+        <div className="flex mt-4 mb-2">
           <CheckBox
             checked={allowReporting === Decision.ACCEPTED}
             onChange={handleReportingChanged}
           />
-          <h3>Error Reporting</h3>
+          <h3 className="font-bold ml-2">Error Reporting</h3>
         </div>
         <div>
           Not every website code is perfect. We use Sentry.io to collect errors which occur during
@@ -63,45 +63,45 @@ const Info = () => {
           reporting you help us to identify and fix potential errors quickly. More information can
           be found in the privacy policy below.
         </div>
-      </div>
-      <div className="info-box">
-        <h2>Privacy Policy</h2>
-        <p>
+      </Box>
+      <Box>
+        <h2 className="text-xl font-bold">Privacy Policy</h2>
+        <p className="my-4">
           We want you to understand how we collect, use and share information about you when you
           use this website. If you have any questions, please don’t hesitate to contact us at
           hello@referee.app.
         </p>
-        <h3>Overview</h3>
-        <p>
-          This website can be used free of charge without registering. We use your personal
+        <h3 className="font-bold">Overview</h3>
+        <p className="my-4">
+        This website can be used free of charge without registering. We use your personal
           information only for providing and improving this website. We only process usage data
           which can include your IP address, browser type and version, operating system, length of
           visit, page views and website navigation paths.
         </p>
-        <h3>Purpose of data collection</h3>
-        <p>
+        <h3 className="font-bold">Purpose of data collection</h3>
+        <p className="my-4">
           We collect this information to analyze how users interact with our website and how many
           users access our website.
         </p>
-        <h3>Cookies</h3>
-        <p>
-          Cookies are files with small amount of data, which may include an anonymous unique
+        <h3 className="font-bold">Cookies</h3>
+        <p className="my-4">
+        Cookies are files with small amount of data, which may include an anonymous unique
           identifier. Cookies are sent to your browser from a web site and stored on your
           computer&apos;s hard drive.
         </p>
-        <p>
+        <p className="my-4">
           We use cookies to gather statistics on the use of our services and evaluate them in
           order to optimize your experience. These cookies allow us to automatically recognize that
           you have visited our site before. The cookies are automatically deleted after a
           pre-defined period. Most browsers accept cookies automatically. However, you can disable
           cookies on your browser or choose to be notified when a new cookie is created.
         </p>
-        <p>
+        <p className="my-4">
           You can use this site without cookies.
         </p>
-        <h3>Sentry</h3>
-        <p>
-          We are constantly improving and developing our website to provide our users with the best
+        <h3 className="font-bold">Sentry</h3>
+        <p className="my-4">
+        We are constantly improving and developing our website to provide our users with the best
           possible user experience. However, not all malfunctions, such as those caused by
           programming errors, can be reliably ruled out from the start. Therefore, we use Sentry, an
           error tracking tool from Functional Software Inc., 132 Hawthorne St, San Francisco, CA
@@ -114,7 +114,7 @@ const Info = () => {
           <li>The IP address of the device</li>
           <li>Date and time of error</li>
         </ul>
-        <p>
+        <p className="my-4">
           There is no explicit analysis for advertising purposes in this process. The information is
           collected anonymously, is not used for personal reasons, and is subsequently deleted. This
           analysis helps us to continuously improve our website and fix hidden code errors. Such
@@ -122,12 +122,12 @@ const Info = () => {
           and analyzing errors. For more information on Sentry’s data processing and how it works,
           see Sentry’s privacy policy: https://sentry.io/privacy/
         </p>
-        <p>
+        <p className="my-4">
           You can object to the delivery of data to Sentry at any time by disabling the checkbox
           above.
         </p>
-        <h3>Google Analytics</h3>
-        <p>
+        <h3 className="font-bold">Google Analytics</h3>
+        <p className="my-4">
           For the continuous improvement of our Services we use the web analytics service Google
           Analytics of Google Inc., 1600 Amphitheatre Parkway, Mountain View, CA 94043, USA
           (hereinafter “Google”). Using cookies, Google creates pseudonymized user profiles. The
@@ -140,7 +140,7 @@ const Info = () => {
           <li>Host name of the accessing computer (IP address)</li>
           <li>Time of the server request</li>
         </ul>
-        <p>
+        <p className="my-4">
           This information is sent to a Google server in the U.S. and stored there. The
           information is used to evaluate the use of our services, to compile reports on the
           activities, and to provide other related services for purposes of market research and
@@ -149,14 +149,14 @@ const Info = () => {
           will your IP address be merged with any other Google data. The IP addresses are
           anonymized so that assignment is not possible (IP masking).
         </p>
-        <p>
+        <p className="my-4">
           We only send data to Google after you have explicitly agreed to it by clicking on
           “Accept” in the cookie banner or checking the checkbox above. You can find out more
           about Google&apos;s use of information by
           visiting https://www.google.com/policies/privacy/partners/
         </p>
-        <h3>Recipients outside the EU</h3>
-        <p>
+        <h3 className="font-bold">Recipients outside the EU</h3>
+        <p className="my-4">
           As indicated above, data may also be sent to recipients located outside the European
           Union or the European Economic Area. This applies in particular to the aforementioned
           processing of analysis technologies, which can result in data transmission to the
@@ -164,12 +164,12 @@ const Info = () => {
           the US. We only work with those service providers who are certified by the EU-US
           Privacy Shield.
         </p>
-        <h3>Your Rights</h3>
-        <p>
-          In addition to the right at any time to withdraw any consent you have given us, you
+        <h3 className="font-bold">Your Rights</h3>
+        <p className="my-4">
+        In addition to the right at any time to withdraw any consent you have given us, you
           are also entitled to the following if the respective legal conditions are met:
         </p>
-        <ul>
+        <ul className="my-4">
           <li>the right to access - you can ask for copies of your personal data</li>
           <li>
             the right to rectification - you can ask us to rectify inaccurate personal data
@@ -197,13 +197,13 @@ const Info = () => {
             processing of your personal data is consent, you can withdraw that consent.
           </li>
         </ul>
-        <h3>Data Security</h3>
-        <p>
+        <h3 className="font-bold">Data Security</h3>
+        <p className="my-4">
           We apply the highest standards to data security for our infrastructure and
           the processing of your data. For example, we use protection mechanisms for
           computers such as firewalls and data encryption.
         </p>
-        <p>
+        <p className="my-4">
           All personal data sent by you is also transmitted using the generally
           accepted and secure SSL (Secure Socket Layer) standard. SSL is a secure and
           proven standard, e.g. it is also used for online banking. You will recognize
@@ -211,7 +211,7 @@ const Info = () => {
           (i.e. https: // …) in the address bar of your browser, or with the lock
           icon at the bottom of the browser.
         </p>
-      </div>
+      </Box>
     </div>
   );
 };