+++ src/components/Item.tsx
@@ -1,27 +1,36 @@
-import React, { FunctionComponent, ReactNode } from "react";
+import React, { PropsWithChildren, ReactNode } from "react";
+import classNames from "classnames";
 
-export interface ItemProps {
-  code: string;
-  className?: string;
-  onClick?: () => void;
+export interface ItemProps<T> {
+  code: T; // eslint-disable-line react/no-unused-prop-types
+  type?: "header" | "regular"
   selected?: boolean;
+  onClick?: () => void;
   children: ReactNode;
+  className?: string;
 }
 
-const Item: FunctionComponent<ItemProps> = ({
+function Item<T>({
   selected,
+  type,
   children,
   onClick,
-  className,
-}) => (
-  <li
-    className={className}
-    role="option"
-    aria-selected={selected}
-    onClick={onClick}
-  >
-    {children}
-  </li>
-);
+  className: additionalClassName,
+}: PropsWithChildren<ItemProps<T>>) {
+  const className = classNames("flex p-1 content-center", {
+    "aria-selected:bg-blue-500 hover:bg-blue-200": type === "header",
+    "aria-selected:bg-grey-300 hover:bg-grey-400": type === "regular",
+  }, additionalClassName);
+  return (
+    <li
+      className={className}
+      role="option"
+      aria-selected={selected}
+      onClick={onClick}
+    >
+      {children}
+    </li>
+  );
+}
 
 export default Item;