+++ src/components/stats/Stats.tsx
@@ -1,12 +1,12 @@
 /* eslint-disable no-param-reassign, no-mixed-operators */
-import React, { FunctionComponent, useMemo } from "react";
+import React, {FunctionComponent, useMemo, useState} from "react";
 import { useTranslation } from "react-i18next";
 import { faCheck, faPercent, faQuestion } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useRulesTestData } from "../../context/TestDataContext";
 import Question from "../../model/Question";
-import "./Stats.css";
 import Rule from "./Rule";
+import Box from "../Box";
 
 type OrderedData = {
   [rule: string]: {
@@ -17,8 +17,11 @@ type OrderedData = {
 };
 
 const Stats: FunctionComponent = () => {
-  const { asked, correct, data } = useRulesTestData();
+  const {
+    asked, correct, data, resetStats,
+  } = useRulesTestData();
   const { t } = useTranslation();
+  const [rerender, setRerender] = useState(0);
 
   const percent = asked ? Math.round(100 / asked * correct) : 0;
   const orderedData = useMemo(() => Object.values(data).reduce<OrderedData>((prev, question) => {
@@ -38,7 +41,12 @@ const Stats: FunctionComponent = () => {
     };
 
     return prev;
-  }, {}), [data]);
+  }, {}), [data, rerender]);
+
+  const handleReset = async () => {
+    await resetStats();
+    setRerender(rerender + 1);
+  };
 
   const rules = Object.keys(orderedData).map((id: string) => {
     const ruleData = orderedData[id];
@@ -54,22 +62,25 @@ const Stats: FunctionComponent = () => {
   });
 
   return (
-    <div id="stats">
-      <div id="stats-overall">
-        <h2>{t("stats.overall")}</h2>
-        <div id="stats-asked">
-          <FontAwesomeIcon icon={faQuestion} size="lg" />
+    <div className="overflow-auto text-xl">
+      <Box>
+        <h2 className="font-bold">{t("stats.overall")}</h2>
+        <div className="text-blue-300 mt-2">
+          <FontAwesomeIcon className="inline-block mr-4 w-5" icon={faQuestion} size="lg" />
           {`${asked} ${t("stats.asked")}`}
         </div>
-        <div id="stats-correct">
-          <FontAwesomeIcon icon={faCheck} size="lg" />
+        <div className="text-green-300 mt-2">
+          <FontAwesomeIcon className="inline-block mr-4 w-5" icon={faCheck} size="lg" />
           {`${correct} ${t("stats.correct")}`}
         </div>
-        <div id="stats-percent">
-          <FontAwesomeIcon icon={faPercent} size="lg" />
+        <div className="text-orange-300 mt-2">
+          <FontAwesomeIcon className="inline-block mr-4 w-5" icon={faPercent} size="lg" />
           {`${percent}%`}
         </div>
-      </div>
+        <button className="border mt-4 w-full md:w-60 rounded-md active:bg-grey-300" onClick={handleReset}>
+          {t("stats.reset")}
+        </button>
+      </Box>
       {rules}
     </div>
   );