+++ src/context/TestDataContext.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import invariant from "ts-invariant";
 import Question from "../model/Question";
 import { ITestResponse } from "../model";
+import { AVAILABLE_LANGUAGES } from "../model/TestDataManager";
 
 export interface TestDataContextValue {
   checkAnswers?: (options: string[]) => Promise<ITestResponse>;
@@ -12,6 +13,10 @@ export interface TestDataContextValue {
   checked: string[];
   data: {[id: string]: Question };
   reveal: boolean;
+  languageMissing: boolean;
+  version: keyof typeof AVAILABLE_LANGUAGES;
+  setVersion?: (version: keyof typeof AVAILABLE_LANGUAGES) => void;
+  resetStats: () => Promise<void>;
 }
 
 let TestDataContext: React.Context<TestDataContextValue>;
@@ -23,7 +28,10 @@ export function getTestDataContext() {
       correct: 0,
       data: {},
       checked: [],
-      reveal: false
+      reveal: false,
+      languageMissing: false,
+      version: "ihf_05_2024",
+      resetStats: async () => {},
     });
   }
   return TestDataContext;